Refractoring and Code Polishing - 12/12/07 (See description)
Function calls have been modified to send parameters via reference, added const control, correct variable cast and other general polishing.

+++ ModuleTime.cpp
@@ -23,8 +23,8 @@ bool ModuleTime::Init()
 update_status ModuleTime::PreUpdate()
 {
 	Uint32 tick = SDL_GetTicks();
-	deltaTime = (tick - lastDeltaTime) / 1000.0f;
-	lastDeltaTime = tick;
+	deltaTime = ((float)tick - lastDeltaTime) / 1000.0f;
+	lastDeltaTime = (float)tick;
 
 	return UPDATE_CONTINUE;
 }
